% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point-trace.R
\name{create_trace_layers}
\alias{create_trace_layers}
\title{Create geom_*_trace layers}
\usage{
create_trace_layers(
  mapping,
  data,
  stat,
  geom,
  position,
  show.legend,
  inherit.aes,
  params,
  trace_position,
  background_params,
  trans_fn,
  allow_bottom = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{geom}{The geometric object to use to display the data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{params}{Additional parameters to pass to the geom and stat.}

\item{trace_position}{Specifies which groups of data points should be
outlined. Can be 'all', 'bottom', or a predicate to use for filtering
data. If 'all', the default, every group plotted will be outlined, if
'bottom', only the bottom most layer of points will be outlined. A
subset of data points can be outlined by passing a predicate. This must
evaluate to `TRUE` or `FALSE` within the context of the input data.}

\item{background_params}{Named list specifying aesthetic parameters to use
for background points when a predicate is passed to `trace_position`.}

\item{trans_fn}{Function to use for transforming data when predicate is
passed to trace_position. Must accept three arguments: dat, data to transform;
ex, expression to use for transforming data; inv, should the expression be
negated, this allows the inverse data points to be transformed.}

\item{allow_bottom}{Should 'bottom' be allowed as an argument for trace_position?}
}
\description{
Create geom_*_trace layers
}
