% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-line-trace.R
\name{geom_path_trace}
\alias{geom_path_trace}
\title{Connect observations}
\usage{
geom_path_trace(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arrow = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Connect observations
}
