% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbody.R
\name{ggbody}
\alias{ggbody}
\title{Return the callstack of a ggproto method as a list}
\usage{
ggbody(method, obj)
}
\arguments{
\item{method}{The method name as a string. Alternatively an expression
that evaluates to the ggproto method in the form of \code{ggproto$method}.}

\item{obj}{The ggproto object. Can be omitted if the method is an expression
in the form of \code{ggproto$method} that evalutes to the ggproto object's method.}
}
\description{
Return the callstack of a ggproto method as a list
}
