% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-offs.R
\name{ggdebug}
\alias{ggdebug}
\alias{ggdebugonce}
\alias{ggundebug}
\title{Debug a ggproto method}
\usage{
ggdebug(method, ...)

ggdebugonce(method, ...)

ggundebug(method, ...)
}
\arguments{
\item{method}{A function or a ggproto method.
The ggproto method may be specified using any of the following forms:
\itemize{
\item \code{ggproto$method}
\item \code{namespace::ggproto$method}
\item \code{namespace:::ggproto$method}
}}

\item{...}{Ignored. Designed for the ease of calling this function by modifying the call to
an earlier \code{{ggtrace}} function in interactive contexts.}
}
\description{
Debug a ggproto method
}
