% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-utils.R
\name{ggeval_silent}
\alias{ggeval_silent}
\title{Simulate printing of a ggplot object without side effects}
\usage{
ggeval_silent(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calls \code{ggplot_gtable(ggplot_build(x))}, which executes all the instructions for plotting \code{x} without
the side-effects from \code{ggplot2:::print.ggplot}, namely:
}
\details{
\itemize{
\item \code{last_plot()} is not overriden
\item \code{grid.draw()} is not called (= nothing gets rendered/drawn)
}
}
