% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-line-trace.R, R/geom-point-trace.R,
%   R/utilities-ggplot2.R
\docType{data}
\name{GeomPathTrace}
\alias{GeomPathTrace}
\alias{GeomPointTrace}
\alias{ggplot2-ggproto}
\title{GeomPointTrace}
\description{
If you are creating a new geom, stat, position, or scale in another package,
you'll need to extend from `ggplot2::Geom`, `ggplot2::Stat`,
`ggplot2::Position`, or `ggplot2::Scale`.
}
\seealso{
ggproto
}
\keyword{datasets}
\keyword{internal}
