% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-path-trace.R, R/geom-point-trace.R
\docType{data}
\name{GeomPathTrace}
\alias{GeomPathTrace}
\alias{GeomLineTrace}
\alias{GeomStepTrace}
\alias{GeomPointTrace}
\title{GeomPathTrace}
\value{
ggproto object
}
\description{
GeomPathTrace
}
\seealso{
\link[ggplot2]{GeomPath}
}
\keyword{datasets}
