% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtrace-package.R
\docType{package}
\name{ggtrace}
\alias{ggtrace}
\alias{ggtrace-package}
\title{ggtrace}
\description{
This package contains extra geoms for \pkg{ggplot2}.
}
\details{
Please see the help pages listed below:

\itemize{
  \item \code{\link{geom_point_trace}}
}

Also see the vignette for more usage examples:

\code{browseVignettes("ggtrace")}

Please report issues and suggest improvements at Github:

\url{https://github.com/rnabioco/ggtrace}
}
\author{
\strong{Maintainer}: Ryan Sheridan \email{ryan.sheridan@cuanschutz.edu} (\href{https://orcid.org/0000-0003-4012-3147}{ORCID})

Other contributors:
\itemize{
  \item RNA Bioscience Initiative [funder]
}

}
\keyword{internal}
