% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-inspect.R
\name{ggtrace_inspect_return}
\alias{ggtrace_inspect_return}
\title{Inspect the return value of a method}
\usage{
ggtrace_inspect_return(x, method, cond = TRUE)
}
\arguments{
\item{x}{A ggplot object}

\item{method}{A function or a ggproto method.
The ggproto method may be specified using any of the following forms:
\itemize{
\item \code{ggproto$method}
\item \code{namespace::ggproto$method}
\item \code{namespace:::ggproto$method}
}}

\item{cond}{When the return value should be replaced}
}
\value{
If nothing is returned, \code{NULL}
}
\description{
Inspect the return value of a method
}
