% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-offs.R
\name{ggtraceback}
\alias{ggtraceback}
\title{Inspect the call stack on enter}
\usage{
ggtraceback(method, ...)
}
\arguments{
\item{method}{A function or a ggproto method.
The ggproto method may be specified using any of the following forms:
\itemize{
\item \code{ggproto$method}
\item \code{namespace::ggproto$method}
\item \code{namespace:::ggproto$method}
}}

\item{...}{
  Arguments passed on to \code{\link[rlang:trace_back]{rlang::trace_back}}
  \describe{
    \item{\code{top}}{The first frame environment to be included in the
backtrace. This becomes the top of the backtrace tree and
represents the oldest call in the backtrace.

This is needed in particular when you call \code{trace_back()}
indirectly or from a larger context, for example in tests or
inside an RMarkdown document where you don't want all of the
knitr evaluation mechanisms to appear in the backtrace.

If not supplied, the \code{rlang_trace_top_env} global option is
consulted. This makes it possible to trim the embedding context
for all backtraces created while the option is set. If knitr is
in progress, the default value for this option is
\code{knitr::knit_global()} so that the knitr context is trimmed out
of backtraces.}
    \item{\code{bottom}}{The last frame environment to be included in the
backtrace. This becomes the rightmost leaf of the backtrace tree
and represents the youngest call in the backtrace.

Set this when you would like to capture a backtrace without the
capture context.

Can also be an integer that will be passed to \code{\link[rlang:caller_env]{caller_env()}}.}
  }}
}
\description{
Inspect the call stack on enter
}
\details{
Calls \code{rlang::trace_back()} upon first entering the function or method. Note that \code{ggtraceback()}
only ever prints the back trace once because it calls \code{gguntrace()} after the first time it's
triggered. For more complex inspections of the call stack, \code{ggedit()} is recommended.
}
\note{
The output of \code{rlang::trace_back()} is not logged to the tracedumps (\code{last_ggtrace()} and
\code{global_ggtrace()})
}
\keyword{internal}
