% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gguntrace.R
\name{gguntrace}
\alias{gguntrace}
\title{Untrace a ggproto method}
\usage{
gguntrace(method, ...)
}
\arguments{
\item{method}{An expression that evaluates to the ggproto method.
This may be specified using any of the following forms:\preformatted{- `ggproto$method`

- `namespace::ggproto$method`

- `namespace:::ggproto$method`
}}

\item{...}{Ignored. Designed for the ease of calling this function by modifying the call to other \code{{ggtrace}} functions.}
}
\description{
For explicitly calling \code{untrace()} on a ggproto object.
}
\section{Gotchas}{

\itemize{
\item If you try to untrace a method that is not currently being traced,
you will get a \verb{could not find function "<method>"} error. This is
in contrast to if you tried to untrace simple functions like in
\code{untrace(mean)}. Future updates to \code{gguntrace()} will improve
handling of this default behavior from \code{base::trace()}.
}
}

\examples{
\dontrun{
# jitter_plot <- ggplot(diamonds[1:1000,], aes(cut, depth)) +
#   geom_point(position = position_jitter(width = 0.2, seed = 2021))
# ggedit(PositionJitter$compute_layer)
# jitter_plot
# gguntrace(PositionJitter$compute_layer)
# jitter_plot
}
}
\seealso{
\code{\link[=ggtrace]{ggtrace()}}, \code{\link[=ggedit]{ggedit()}}
}
