% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_traced.R
\name{is_traced}
\alias{is_traced}
\title{Check if a method is being traced}
\usage{
is_traced(method)
}
\arguments{
\item{method}{An expression that evaluates to the ggproto method.
This may be specified using any of the following forms:\preformatted{- `ggproto$method`

- `namespace::ggproto$method`

- `namespace:::ggproto$method`
}}
}
\value{
logical
}
\description{
Check if a method is being traced
}
\examples{
\dontrun{
ggtrace(Stat$compute_layer, 1)
is_traced(Stat$compute_layer)
gguntrace(Stat$compute_layer)
is_traced(Stat$compute_layer)
}
}
