% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_ggtrace.R
\name{last_ggtrace}
\alias{last_ggtrace}
\title{Retrieve the trace dump created by the last ggtrace}
\usage{
last_ggtrace()
}
\description{
Retrieve the trace dump created by the last ggtrace
}
\keyword{internal}
