% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last-errorcontext.R
\name{last_layer_errorcontext}
\alias{last_layer_errorcontext}
\alias{last_sublayer_errorcontext}
\title{Get the internal context of the last (sub-)layer error}
\usage{
last_layer_errorcontext(reprint_error = FALSE, ggtrace_notes = TRUE)

last_sublayer_errorcontext(reprint_error = FALSE, ggtrace_notes = TRUE)
}
\arguments{
\item{reprint_error}{Re-prints the original error message to the console. Defaults to \code{FALSE}.}

\item{ggtrace_notes}{Prints the \code{ggtrace_inspect_args()} call used to inspect the error context. Defaults to \code{TRUE}.}
}
\value{
An dynamically constructed and evaluated call to \code{\link[=ggtrace_inspect_args]{ggtrace_inspect_args()}}.
Prioritizes showing the state of layer data whenever possible (by extracting the \code{data} argument).
}
\description{
\itemize{
\item \code{last_layer_errorcontext()} returns the error context at the level of the \code{Layer} ggproto.
\item \code{last_sublayer_errorcontext()} (EXPERIMENTAL) returns the error context at the sub-\code{Layer} level (e.g., \code{Stat} or \code{Geom}).
}
}
\section{Scope}{

These functions can only retrieve information from errors propagating from
\code{Layer} ggproto methods. In non-technical terms, they only work for errors with a
\strong{"Error occured in the Nth layer"} message (as of \code{{ggplot2}} >= 3.4.0).

The scope of \code{last_sublayer_errorcontext()} is narrower, since not all \code{Layer} methods call
a sub-Layer method. This function is intended for developers - in most cases users can
get all the information necessary to debug layer code from \code{last_layer_errorcontext()}
(there are only so many ways to break a ggplot from user-facing code).
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)
erroring_barplot1 <- ggplot(mtcars, aes(mpg, hp)) +
  stat_summary() +
  geom_bar()

# Render to trigger error
erroring_barplot1

# Both return the same snapshot of layer data
# but at different levels of specificity
last_layer_errorcontext()
last_sublayer_errorcontext()

erroring_barplot2 <- ggplot(mtcars, aes(mpg, hp)) +
  stat_summary() +
  geom_bar(aes(y = c(1, 2)))
erroring_barplot2

# This works:
last_layer_errorcontext()
# This doesn't: there's no sub-layer ggproto involved in this error
last_sublayer_errorcontext()

library(ggforce)
erroring_sina <- ggplot(mtcars, aes(mpg)) +
  geom_bar() +
  geom_sina()
erroring_sina

# The two return different snapshots of layer data here -
# see `ggplot2:::Layer$compute_statistic` for why.
last_layer_errorcontext()
last_sublayer_errorcontext()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=ggtrace_inspect_on_error]{ggtrace_inspect_on_error()}}
}
