% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_trace_colour_discrete}
\alias{scale_trace_colour_discrete}
\title{Discrete trace color scales}
\usage{
scale_trace_colour_discrete(..., type = getOption("ggplot2.discrete.colour"))
}
\arguments{
\item{...}{Additional parameters passed on to the scale type,}

\item{type}{One of the following:
* A character vector of color codes. The codes are used for a 'manual' color
scale as long as the number of codes exceeds the number of data levels
(if there are more levels than codes, [scale_color_hue()]/[scale_fill_hue()]
are used to construct the default scale). If this is a named vector, then the color values
will be matched to levels based on the names of the vectors. Data values that
don't match will be set as `na.value`.
* A list of character vectors of color codes. The minimum length vector that exceeds the
number of data levels is chosen for the color scaling. This is useful if you
want to change the color palette based on the number of levels.
* A function that returns a discrete color/fill scale (e.g., [scale_fill_hue()],
[scale_fill_brewer()], etc).}
}
\description{
The default discrete color scale. Defaults to [scale_fill_hue()]/[scale_fill_brewer()]
unless `type` (which defaults to the `ggplot2.discrete.fill`/`ggplot2.discrete.color` options)
is specified.
}
