% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{EuStockMarkets in long format}
\format{
A tibble with 74440 rows and 3 variables
}
\usage{
stocks
}
\description{
EuStockMarkets in long format
}
\keyword{datasets}
