% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{workflow-function-aliases}
\alias{workflow-function-aliases}
\title{Shorthand aliases for workflow functions}
\description{
These functions are identical to their long-form counterparts suffixed
by \code{ggtrace_}.

Inspect:
\itemize{
\item \code{\link[=inspect_n]{inspect_n()}}
\item \code{\link[=inspect_which]{inspect_which()}}
\item \code{\link[=inspect_vars]{inspect_vars()}}
\item \code{\link[=inspect_args]{inspect_args()}}
\item \code{\link[=inspect_return]{inspect_return()}}
\item \code{\link[=inspect_on_error]{inspect_on_error()}}
}

Capture:
\itemize{
\item \code{\link[=capture_fn]{capture_fn()}}
\item \code{\link[=capture_env]{capture_env()}}
}

Highjack:
\itemize{
\item \code{\link[=highjack_args]{highjack_args()}}
\item \code{\link[=highjack_return]{highjack_return()}}
}
}
\keyword{internal}
