% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point-trace.R
\name{create_trace_layers}
\alias{create_trace_layers}
\title{Create geom_*_trace layers}
\usage{
create_trace_layers(
  mapping,
  data,
  stat,
  geom,
  position,
  show.legend,
  inherit.aes,
  params,
  trace_position,
  background_color,
  allow_bottom = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() to use for generating layers.}

\item{data}{The data to be displayed in these layers.}

\item{stat}{Statistical transformation to use for these layers.}

\item{geom}{The geometric object to use to display the data.}

\item{position}{Position adjustment.}

\item{show.legend}{Should these layers be included in the legend?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics.}

\item{params}{Additional parameters to pass to the geom and stat.}

\item{trace_position}{Specifies which groups of data points should be
outlined. Can be 'all', 'bottom', or a predicate to use for filtering
data. If 'all', the default, every group plotted will be outlined, if
'bottom', only the bottom most layer of points will be outlined. A
subset of data points can be outlined by passing a predicate. This must
evaluate to `TRUE` or `FALSE` within the context of the input data.}

\item{background_color}{Color to use for background points when a predicate
is passed to `trace_position`. If NULL, the original fill color will be
used.}

\item{allow_bottom}{Should 'bottom' be allowed as an argument for trace_position?}
}
\description{
Create geom_*_trace layers
}
