% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_method}
\alias{get_method}
\title{Return ggproto methods as functions}
\usage{
get_method(method)
}
\arguments{
\item{method}{A function or a ggproto method.
The ggproto method may be specified using any of the following forms:
\itemize{
\item \code{ggproto$method}
\item \code{namespace::ggproto$method}
\item \code{namespace:::ggproto$method}
}}
}
\value{
Function
}
\description{
Return ggproto methods as functions
}
\examples{
library(ggplot2)

# Uninformative
StatCount$compute_group
formals(StatCount$compute_group)
body(StatCount$compute_group)

# Errors
# get(StatCount$compute_group)

# Informative
get_method(StatCount$compute_group)
formals(get_method(StatCount$compute_group))
body(get_method(StatCount$compute_group))
}
