% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbody.R
\name{ggbody}
\alias{ggbody}
\title{Retrieve the body of a ggproto method as a list}
\usage{
ggbody(method, inherit = FALSE)
}
\arguments{
\item{method}{An expression that evaluates to the ggproto method.
This may be specified using any of the following forms:\preformatted{- `ggproto$method`
- `namespace::ggproto$method`
- `namespace:::ggproto$method`
}}

\item{inherit}{Whether the method should be returned from its closest parent.
Defaults to \code{FALSE}.}
}
\description{
Retrieve the body of a ggproto method as a list
}
\details{
For interactive uses, using the short form is recommended. For
programmatic uses, \code{as.list(body(get("method", ggproto)))} is recommended.

The \code{get("method", ggproto)} syntax is the long form of \code{ggproto$method} with
subtle but important differences;
\itemize{
\item For example, this works: \code{debugonce(get("compute_group", StatCount))}
\item But this fails to insert a break point: \code{debugonce(StatCount$compute_group)}
}
}
\examples{
\dontrun{
library(ggplot2)

ggbody(StatCount$compute_group)

# Works for ggproto in extension packages

ggbody(ggforce::StatDelaunaySegment$compute_group)

library(ggforce)
ggbody(StatBezier$compute_panel)

# `inherit = TRUE` will return the method from the closest parent

# ERRORS:
# ggbody(StatBoxplot$compute_panel)
ggbody(StatBoxplot$compute_panel, inherit = TRUE)
ggbody(Stat$compute_panel)

# Navigating complex inheritance
class(GeomArcBar)
invisible(ggbody(GeomArcBar$default_aes, inherit = TRUE)) # self
invisible(ggbody(GeomArcBar$draw_panel, inherit = TRUE))  # parent
invisible(ggbody(GeomArcBar$draw_key, inherit = TRUE))    # grandparent
invisible(ggbody(GeomArcBar$draw_group, inherit = TRUE))  # top-level

}
}
