% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbody.R
\name{ggbody}
\alias{ggbody}
\title{Retrieve the body of a ggproto method as a list}
\usage{
ggbody(method, obj, inherit = FALSE)
}
\arguments{
\item{method}{The method name as a string. Alternatively an expression
that evaluates to the ggproto method in the form of \code{ggproto$method}.}

\item{obj}{The ggproto object. Can be omitted if the method is an expression
in the form of \code{ggproto$method} that evalutes to the ggproto object's method.}

\item{inherit}{Whether the method should be returned from its closest parent.
Defaults to \code{FALSE}.}
}
\description{
Retrieve the body of a ggproto method as a list
}
\details{
Despite the convenience of the short form which \code{ggbody()} also works with,
the long form of retrieving the ggproto method by specifying both the method and
the object separately exists for consistency with other ways of inspecting
ggproto methods.
\itemize{
\item For example, this works: \code{debugonce(get("compute_group", StatCount))}
\item But this fails to insert a break point: \code{debugonce(StatCount$compute_group)}
}
}
\examples{
\dontrun{
library(ggplot2)

# Methods can be specified in both long form and short form

longform <- ggbody("compute_group", StatCount)
longform

shortform <- ggbody(StatCount$compute_group)
shortform

identical(longform, shortform)

# Works for ggproto in extension packages

library(ggforce)
ggbody(StatBezier$compute_panel)

# `inherit = TRUE` will return the method from the closest parent

ggbody(StatBoxplot$compute_panel, inherit = TRUE)
ggbody(Stat$compute_panel)

}
}
