% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggedit.R
\name{ggedit}
\alias{ggedit}
\title{Interactively edit the source code of a ggproto method}
\usage{
ggedit(method, ...)
}
\arguments{
\item{method}{An expression that evaluates to the ggproto method.
This may be specified using any of the following forms:\preformatted{- `ggproto$method`

- `namespace::ggproto$method`

- `namespace:::ggproto$method`
}}

\item{...}{Ignored. Designed for the ease of calling this function by modifying the call to other \code{{ggtrace}} functions.}
}
\description{
Interactively edit the source code of a ggproto method
}
\details{
Like \code{base::trace()}, the edit is in place until \code{untrace()} is called.
To \code{untrace()} a ggproto method, the syntax is \code{untrace(what = "method", where = obj)}
}
\examples{
\dontrun{
# jitter_plot <- ggplot(diamonds[1:1000,], aes(cut, depth)) +
#   geom_point(position = position_jitter(width = 0.2, seed = 2021))
# ggedit(PositionJitter$compute_layer)
# # < interactively modify the method's source code in text editor >
# jitter_plot # Edit is in place
# gguntrace(PositionJitter$compute_layer)
# # Or untrace(what = "compute_layer", where = PositionJitter)
# jitter_plot # Edit is removed
}
}
\seealso{
\code{\link[=gguntrace]{gguntrace()}}
}
