% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-utils.R
\name{ggeval_silent}
\alias{ggeval_silent}
\alias{ggdraw_silent}
\title{Simple evaluation and printing of a ggplot object}
\usage{
ggeval_silent(x)

ggdraw_silent(x)
}
\arguments{
\item{x}{A ggplot}
}
\value{
A gtable
}
\description{
\code{ggeval_silent()} evaluates the ggplot object into a \verb{<gtable>} grob and \code{ggdraw_silent()}
renders the grob with \code{{grid}}. Combined, these two functions simulate \code{ggplot2:::print.ggplot}
without overriding \code{last_plot()}.
}
