% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtrace-package.R
\docType{package}
\name{ggtrace}
\alias{ggtrace}
\alias{ggtrace-package}
\title{ggtrace}
\description{
This package provides \pkg{ggplot2} geoms that allow groups of data points to be
outlined or highlighted for emphasis. This is particularly useful when
working with dense datasets that are prone to overplotting.
}
\details{
See the vignette for usage examples:

\code{browseVignettes("ggtrace")}
}
\seealso{
Report issues and suggest improvements at \url{https://github.com/rnabioco/ggtrace/issues}
}
\author{
Ryan Sheridan \email{ryan.sheridan@cuanschutz.edu}
}
\keyword{internal}
