% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-capture.R
\name{ggtrace_capture_env}
\alias{ggtrace_capture_env}
\title{Capture a snapshot of a method's execution environment}
\usage{
ggtrace_capture_env(x, ..., at = -1L)
}
\arguments{
\item{x}{A ggplot object}

\item{...}{Passed to \code{ggtrace()}. The \code{method} to capture should be specified here.}

\item{at}{The position in the method body when the environment should be captured.
Defaults to \code{-1L}, which is right before the method returns.}
}
\value{
An environment
}
\description{
Capture a snapshot of a method's execution environment
}
\examples{
library(ggplot2)

# Example from https://ggplot2.tidyverse.org/reference/aes_eval.html
after_scale_plot <- ggplot(mpg, aes(class, hwy)) +
  geom_boxplot(aes(colour = class, fill = after_scale(alpha(colour, 0.4))))
after_scale_plot

# `after_scale()` is resolved by `Geom$use_defaults` (at Step 6)

before_applying <- ggtrace_capture_env(
  x = after_scale_plot,
  method = Geom$use_defaults,
  at = 1  # To be more specific, do `at = 6`
)
after_applying <- ggtrace_capture_env(
  x = after_scale_plot,
  method = Geom$use_defaults,
  at = -1  # To be more specific, do `at = 7`
)

colnames(before_applying$data)
colnames(after_applying$data)

library(dplyr)

before_applying$data \%>\%
  select(any_of(c("colour", "fill")))
after_applying$data \%>\%
  select(any_of(c("colour", "fill")))

identical(
  before_applying$data \%>\%
    select(any_of(c("colour", "fill"))) \%>\%
    mutate(fill = alpha(colour, 0.4)),       #< after_scale() logic here
  after_applying$data \%>\%
    select(any_of(c("colour", "fill")))
)

}
