% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-highjack.R
\name{ggtrace_highjack_return}
\alias{ggtrace_highjack_return}
\title{Highjack a method's execution and make it return a user-supplied value}
\usage{
ggtrace_highjack_return(
  x,
  method,
  cond = quote(._counter_ == 1),
  value = quote(returnValue()),
  draw = TRUE
)
}
\arguments{
\item{x}{A ggplot object}

\item{method}{A function or a ggproto method.
The ggproto method may be specified using any of the following forms:
\itemize{
\item \code{ggproto$method}
\item \code{namespace::ggproto$method}
\item \code{namespace:::ggproto$method}
}}

\item{cond}{When the return value should be replaced.}

\item{value}{What the method should return instead}

\item{draw}{Whether to draw the modified graphical output from evaluating \code{x}.
Defaults to \code{TRUE}.}
}
\value{
A gtable object with class \verb{<ggtrace_highjacked>}
}
\description{
Highjack a method's execution and make it return a user-supplied value
}
