% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows-highjack.R
\name{ggtrace_highjack_return}
\alias{ggtrace_highjack_return}
\title{Highjack the return value of a method}
\usage{
ggtrace_highjack_return(x, method, value, cond = TRUE, draw = TRUE)
}
\arguments{
\item{x}{A ggplot object}

\item{method}{A function or a ggproto method.
The ggproto method may be specified using any of the following forms:
\itemize{
\item \code{ggproto$method}
\item \code{namespace::ggproto$method}
\item \code{namespace:::ggproto$method}
}}

\item{value}{Value for the method to return instead when it is called}

\item{cond}{When the return value should be replaced}

\item{draw}{Whether to draw the modified plot}
}
\value{
A gtable object with class \verb{<ggtrace_highjacked>}
}
\description{
Highjack the return value of a method
}
