% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gguntrace.R
\name{gguntrace}
\alias{gguntrace}
\title{Untrace a ggproto method}
\usage{
gguntrace(method, obj)
}
\arguments{
\item{method}{The method name as a string. Alternatively an expression
that evaluates to the ggproto method in the form of \code{ggproto$method}.}

\item{obj}{The ggproto object. Can be omitted if the method is an expression
in the form of \code{ggproto$method} that evalutes to the ggproto object's method.}
}
\description{
For explicitly calling \code{untrace()} on a ggproto object.
}
\section{Gotchas}{

\itemize{
\item If you try to untrace a method that is not currently being traced,
you will get a \verb{could not find function "<method>"} error. This is
in contrast to if you tried to untrace simple functions like in
\code{untrace(mean)}. Future updates to \code{gguntrace()} will improve
handling of this default behavior from \code{base::trace()}.
}
}

\examples{
\dontrun{
# jitter_plot <- ggplot(diamonds[1:1000,], aes(cut, depth)) +
#   geom_point(position = position_jitter(width = 0.2, seed = 2021))
# ggedit(PositionJitter$compute_layer)
# jitter_plot
# gguntrace(PositionJitter$compute_layer)
# jitter_plot
}
}
\seealso{
\code{\link[=ggtrace]{ggtrace()}}, \code{\link[=ggedit]{ggedit()}}
}
