% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gguntrace.R
\name{gguntrace}
\alias{gguntrace}
\title{Untrace a ggproto method}
\usage{
gguntrace(method, ...)
}
\arguments{
\item{method}{A function or a ggproto method.
The ggproto method may be specified using any of the following forms:
\itemize{
\item \code{ggproto$method}
\item \code{namespace::ggproto$method}
\item \code{namespace:::ggproto$method}
}}

\item{...}{Ignored. Designed for the ease of calling this function by modifying the call to
an earlier \code{{ggtrace}} function in interactive contexts.}
}
\description{
For explicitly calling \code{untrace()} on a ggproto object.
}
\details{
Unlike \code{base::untrace()}, there is no adverse side effect to repeatedly calling \code{gguntrace()}
on a ggproto method. \code{gguntrace()} will only throw an error if the method cannot be found.

If the method is valid, \code{gguntrace()} will do one of two things:
\itemize{
\item Inform that it has successfully removed the trace (after untracing)
\item Inform that the there isn't an existing trace (after doing nothing)
}
}
\examples{
library(ggplot2)

gguntrace(Stat$compute_layer)

is_traced(Stat$compute_layer)

ggtrace(Stat$compute_layer, 1)

is_traced(Stat$compute_layer)

gguntrace(Stat$compute_layer)

is_traced(Stat$compute_layer)

gguntrace(Stat$compute_layer)

}
\seealso{
\code{\link[=ggtrace]{ggtrace()}}, \code{\link[=ggedit]{ggedit()}}
}
