% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_traced.R
\name{is_traced}
\alias{is_traced}
\title{Check if a method is being traced}
\usage{
is_traced(method)
}
\arguments{
\item{method}{A function or a ggproto method.
The ggproto method may be specified using any of the following forms:
\itemize{
\item \code{ggproto$method}
\item \code{namespace::ggproto$method}
\item \code{namespace:::ggproto$method}
}}
}
\value{
logical
}
\description{
Check if a method is being traced
}
\examples{
library(ggplot2)

gguntrace(Stat$compute_layer)

is_traced(Stat$compute_layer)

ggtrace(Stat$compute_layer, 1)

is_traced(Stat$compute_layer)

gguntrace(Stat$compute_layer)

is_traced(Stat$compute_layer)

gguntrace(Stat$compute_layer)

}
