% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracedump.R
\name{last_ggtrace}
\alias{last_ggtrace}
\alias{clear_last_ggtrace}
\title{Retrieve the last trace dump created by \code{ggtrace()}}
\usage{
last_ggtrace()

clear_last_ggtrace()
}
\value{
List
}
\description{
\code{last_ggtrace()} retrieves the last trace dump created by \code{ggtrace()} - i.e., from the last time
the trace has been triggered.
}
\details{
A \strong{tracedump} refers to the output of an expression evaluated inside a traced function when
a trace is triggered. In the context of \code{ggtrace()}, the tracedump is a list that is as long as the number of
step-expression pairs, where each element is the output of each expression (\code{trace_exprs}) evaluated at each
step (\code{trace_steps}) inside the execution environment of the function or ggproto method.

\code{clear_last_ggtrace()} sets the value of \code{last_ggtrace()} to \code{NULL} and returns \code{NULL}.
}
\examples{
library(ggplot2)

ggtrace(StatSmooth$compute_group, trace_steps = -1, trace_exprs = quote(head(prediction)))

ggplot(mtcars, aes(mpg, hp)) + geom_point() + geom_smooth(method = 'lm')

last_ggtrace()

ggtrace(
  StatSmooth$compute_group,
  trace_steps = -1,
  trace_exprs = quote(prediction),
  verbose = FALSE
)

ggplot(mtcars, aes(mpg, hp)) + geom_point() + geom_smooth(method = 'lm')

head(last_ggtrace()[[1]])

}
\seealso{
\code{\link[=global_ggtrace]{global_ggtrace()}}
}
\keyword{internal}
