% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-trace-alpha.R
\name{scale_trace_alpha}
\alias{scale_trace_alpha}
\alias{scale_trace_alpha_continuous}
\alias{scale_trace_alpha_binned}
\alias{scale_trace_alpha_discrete}
\alias{scale_trace_alpha_ordinal}
\title{Alpha transparency scales}
\usage{
scale_trace_alpha(..., range = c(0.1, 1))

scale_trace_alpha_continuous(..., range = c(0.1, 1))

scale_trace_alpha_binned(..., range = c(0.1, 1))

scale_trace_alpha_discrete(...)

scale_trace_alpha_ordinal(..., range = c(0.1, 1))
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2:continuous_scale]{continuous_scale()}}, \link[ggplot2]{binned_scale},
or \code{\link[ggplot2:discrete_scale]{discrete_scale()}} as appropriate, to control name, limits,
breaks, labels and so forth.}

\item{range}{Output range of alpha values. Must lie between 0 and 1.}
}
\description{
Alpha-transparency scales are not tremendously useful, but can be a
convenient way to visually down-weight less important observations.
scale_alpha() is an alias for scale_alpha_continuous() since that is the
most common use of alpha, and it saves a bit of typing.
}
