% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublayer-data.R
\name{sublayer-data}
\alias{sublayer-data}
\alias{layer_before_stat}
\alias{layer_after_stat}
\title{Inspect snapshots of sub-layer data}
\usage{
layer_before_stat(x, cond = 1L, error = FALSE, ...)

layer_after_stat(x, cond = 1L, error = FALSE, ...)
}
\arguments{
\item{x}{A ggplot object. If missing, defaults to \code{ggplot2::last_plot()}.}

\item{cond}{Index of the layer to inspect. Defaults to \code{1L}.}

\item{error}{For debugging. If \code{TRUE}, continues inspecting the method until the ggplot errors.}

\item{...}{Unused.}
}
\value{
A dataframe
}
\description{
Inspect snapshots of sub-layer data
}
\keyword{internal}
