% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublayer-data.R
\name{sublayer-data}
\alias{sublayer-data}
\alias{layer_before_stat}
\alias{layer_after_stat}
\alias{layer_before_geom}
\alias{layer_after_scale}
\title{Inspect snapshots of sub-layer data}
\usage{
layer_before_stat(plot, i = 1L, ..., error = FALSE)

layer_after_stat(plot, i = 1L, ..., error = FALSE)

layer_before_geom(plot, i = 1L, ..., error = FALSE)

layer_after_scale(plot, i = 1L, ..., error = FALSE)
}
\arguments{
\item{plot}{A ggplot object. If missing, defaults to \code{ggplot2::last_plot()}.}

\item{i}{Index of the layer to inspect. Defaults to \code{1L}.}

\item{...}{Unused.}

\item{error}{For debugging. If \code{TRUE}, continues inspecting the method until the ggplot errors.}
}
\value{
A dataframe
}
\description{
\code{layer_before_stat()}, \code{layer_after_stat()}, \code{layer_before_geom()}, and \code{layer_after_scale()}
are helper functions that return a snapshot of a layer's data in the internals.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)
ggplot(mpg, aes(displ, class)) +
  geom_boxplot(outlier.shape = NA) +
  geom_text(
    aes(
      label = after_stat(xmax),
      x = stage(displ, after_stat = xmax)
    ),
    stat = "boxplot", hjust = -0.5
  )

# Before Stat snapshot of first layer's data
layer_before_stat()

# After Stat snapshot of first layer's data
layer_after_stat()

# First and second layer's data are identical for those two stages
identical(layer_before_stat(), layer_before_stat(i = 2))
identical(layer_after_stat(), layer_after_stat(i = 2))

# `after_stat()` mappings adds new columns to the second layer's data
# by the time the geom receives the data in the Before Geom stage
layer_before_geom(i = 2)
waldo::compare(layer_before_geom(), layer_before_geom(i = 2))
\dontshow{\}) # examplesIf}
}
\keyword{internal}
